WIP posts filter

+++ src/app/modules/posts/components/post-list/post-list.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { PostsService } from '../../services/posts.service';
 import { Post } from '../../models/Post';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../../../../shared/modals/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-posts-list',
@@ -13,14 +15,23 @@ export class PostsListComponent implements OnInit {
     resize: true,
     initLayout: true
   }
-  constructor(public postsService: PostsService) { }
+  constructor(public postsService: PostsService, public dialog: MatDialog) { }
 
   ngOnInit() {
     this.getPosts();
   }
 
   delete(post: Post) {
-    this.postsService.deletePost(post).subscribe();
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '400px',
+      data: {title: 'Confirm', text:`Do you wish to delete the post ${post.name} ?`}
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result) {        
+        this.postsService.deletePost(post).subscribe();
+      }
+    })
   }
 
   getPosts() {
